Migrate MatrixData to MatrixTemplate for item data models
