+++ module/data/common/_module.mjs
@@ -1,4 +1,3 @@
-import { default as MatrixData } from "./matrix.mjs";
 import {
   AreaMetricData,
   DurationMetricData,
@@ -28,7 +27,6 @@ export {
   DurationMetricData,
   EffectMetricData,
   LanguageConstraintData,
-  MatrixData,
   MetricData,
   NamegiverConstraintData,
   RangeMetricData,
