+++ module/data/item/templates/physical-item.mjs
@@ -2,7 +2,7 @@ import { ItemDataModel } from "../../abstract.mjs";
 import TargetTemplate from "./targeting.mjs";
 import ED4E from "../../../config/_module.mjs";
 import ThreadTemplate from "./threads.mjs";
-import MatrixData from "../../common/matrix.mjs";
+import MatrixTemplate from "./matrix.mjs";
 
 /**
  * Data model template with information on physical items.
@@ -19,8 +19,9 @@ import MatrixData from "../../common/matrix.mjs";
  * @property {number} usableItem.recoveryPropertyValue      recovery type value
  */
 export default class PhysicalItemTemplate extends ItemDataModel.mixin(
+  MatrixTemplate,
   TargetTemplate,
-  ThreadTemplate
+  ThreadTemplate,
 ) {
 
   /** @inheritdoc */
@@ -170,7 +171,6 @@ export default class PhysicalItemTemplate extends ItemDataModel.mixin(
         label:    this.labelKey( "PhysicalItems.itemStatus" ),
         hint:     this.hintKey( "PhysicalItems.itemStatus" )
       } ),
-      matrix: MatrixData.asEmbeddedDataField(),
     } );
   }
